Replace for..of with for.
When using for of, the code generated by babel is not ES5 compatible. It
will break in IE11 and below.

The alternative would be to add babel-polyfill. But that my cause
problems if the project that uses react-coral, also uses babel-polyfill.
You need to import babel polyfill only in the top level project.

Because this seems to be the only place which would required
babel-polyfill, I decided to ditch the for..of.