Remove story that shows how to use selected attribute.
The selected attribute is used internally by that TabList component.
TabList has some similarities with TagList and the Tag component doesn't
have a story for `selected` even if the Tag component supports it.

Someone trying to create a controlled TabList component might think the
following code is valid.

```
<TabList>
   <Tab selected>
      Tab 1
   </Tab>
   <Tab>
      Tab 2
   </Tab>
</TabList>
```

It will work, but the TabList component will be in an uncontrolled
state. The right way to make a controlled TabList component is by using
the `selectedIndex` property.