Fix IE11 closing on scroll for autocomplete
The Autocomplete component has an issue with IE11 where the browser will blur (close) the component when the user attempts to click the scrollbar in the dropdown.

This happens because IE11 fires a blur event when clicking the scrollbar which is a known issue that react-select is suppose to handle.

However, the react-select code checks the ‘Select-menu’ div to see if the click happened on that element or as a child of that element. This works great as long as the ‘Select-menu’ div has the scrollbar contained within it (has a css style of overflow-y:auto or equivalent).

This check doesn’t work though when the ‘Select-menu-outer’ div (which wraps ‘Select-menu’) has the style for the scrollbar, which ‘coral-Autocomplete-overlay’ does.

The fix is either to update the react-select code to also check the ‘Select-menu-outer’ container or update the styles to put the overflow-y styles on the ‘Select-menu’ by moving the ‘coral-Autocomplete-overlay’ class. I elected for the latter.