RSP-532 Pause timer for toast when in focus (#418)
* Pause timer for toast when in focus

* Calling original prop methods

* Review comments for memory leak

* Multi toasts dismiss now

When using a javascript object {}, an object cannot be a key, it will get coerced to the string `[object Object]` or something. If we use a Map instead, we can use objects as keys.

* :shame_face: I left console logs in

* fix lint

* fixing focus bug

If there are multiple toasts on the screen and one of them has focus, we don't want it to lose focus as it's siblings disappear

* order matters

when in tests, it matters if you call setState before finishing the rest of add... because it'll try to access the Map too early

* timeout zero

I'm not sure what the use case is for a timeout of zero, maybe someone can clarify, but this now protects us.
Also, not sure how it could be a null value or anything else that `if(timeout)` would fail now that zero is not allowed.