RSP-786 [Accessibility] Accordion: Update implementation to follow WAI-ARIA 1.1 (#513)
* RSP-786 [Accessibility] Accordion: Update implementation to follow WAI-ARIA 1.1

The accessibility implementation for Accordion is problematic with JAWS screen reader.

Treating Accordion as having role=tablist forces JAWS into Forms mode, which requires the user to press Esc key to switch back to virtual cursor mode in order to read the contents of the expanded panel nested within the tablist, adjacent to the expanded tab. Unfortunately, in a Dialog control, the JAWS can't stop propagation of the Esc key event and the Dialog will close, which is undesirable.

A simple solution is to update to use the [WAI-ARIA 1.1 design pattern](https://www.w3.org/TR/wai-aria-practices/#accordion), where the item header is presented to assistive technology as a button wrapped in a heading element. JAWS won't shift into forms mode when an item header receives focus and the panel content will be easier to access.

This issue was identified while testing development in Adobe Connect, where the Accordion items in a dialog showing keyboard shortcuts expand to display content in a table. The table can't be accessed by exiting forms mode, because the Esc key closes the dialog. Connect engineers were forced to add unnecessary keyboard navigation to the static table as a workaround.