Fix #414 [Accessibility] Slider: should support method for deriving aria-valuetext from value (#415)
* Fix #414 [Accessibility] Slider: should support method for deriving aria-valuetext from value

For accessibility, Slider uses aria-valuetext prop so that the value of the input will be read as a text value rather than as a percent of the difference between the min and max. There may be use cases where being able to convert this value into a string rather than a number is desirable. For example, a Slider could be used to pick a time, where min={0} and max={1440}, for the number of minutes in a day. In this case you would want to express the value as a time string, like "10:30 AM" rather than as a number between 0 and 1400, 630.

* Fix #414 Refactor valueTextFunction to getAriaValueText

Per https://github.com/adobe/react-spectrum/pull/415#discussion_r310844194