RSP-1376 Breadcrumbs isHeading support
It should be possible to use any variant of breadcrumbs as a page heading. Currently, we switch the last breadcrumb to an h1 only in the large size. We should instead change to role='heading' and aria-level={current level} if the `isHeading` prop is set to allow any variant to become a heading. Large size also shouldn't imply heading.

https://excelerate-team.slack.com/archives/CB2S6SRK4/p1573496456133100?thread_ts=1573496456133100&cid=CB2S6SRK4

fix lint errors