RSP-1396 [Accessibility] Asterisk has ariaLabel and role presentation
- Fix gap between icon or necessityLabel

With one child, and isRequired on the LabelBase:

- remove the alt text from the required icon or set aria-hidden={true} to the necessityLabel span element.
- ensure that the child also receives isRequired={true} so that it will communicate required state by adding aria-required="true" to the input.

Otherwise, include alt text on the required icon or don't hide the necessityLabel with aria-hidden.

If the LabelBase does not have isRequired, but isRequired is true for the child component, set isRequired on the LabelBase, and also set necessityIndicator = 'icon' if necessityIndicator is undefined on the LabelBase.

RSP-1396 Use zero-width space to force whitespace before necessity label

RSP-1396 Improve comments regarding aria-hidden on required label or icon

Improve aria-required propagation to using Provider.