RSP-1789 Picker: isRequired state should be included in Label
Picker with isRequired does not report required state, because unlike an input, the Picker's popup button control does not support "required" or "aria-required" attribute. We need a mechanism to allow Label to account for what is being labelled as required.

James Nurthen suggests adding a reference to the necessity indicator element using aria-describedby, but we would only want this for the Picker, at this point.

Simple fix is to add a flag to include the necessity indicator in the accessibility name for the Picker's label.