RSP-1834 [Accessibility][Docs] Popup links are not keyboard accessible in Firefox
Popup links have a valid href, so usePress is not interpreting them as links that that React-Spectrum should handle with the Enter key, even though the default behavior has been prevented with event.preventDefault in and onClick handler.

See: https://github.com/adobe-private/react-spectrum-v3/blob/master/packages/%40react-aria/interactions/src/usePress.ts#L573-L577

One solution is to remove href and set role="link" and tabIndex="0". We also have to make sure that links within the Popup still work, but this may be another use case we need to account for in usePress.