RSP-1746 [Safari/VoiceOver] Listbox options with aria-labelledby announce as "Text"
With Safari and VoiceOver, options using `aria-labelledby` or `aria-label` to label the item by referencing an internal content slot are announced simply as "Text". Fix may be to not use `aria-labelledby` or `aria-describedby` to label the option using slots, and simply allow the option to be labeled by its textContent.