Fix `Function components cannot be given refs`
It doesn't make much sense to apply a ref to menutrigger since it doesn't have dom, it should really go to the action button probably, but it's hard to get to there. This i think makes the most sense, it also gives a place to apply the breadcrumb item class name which previously wasn't applied because menutrigger doesn't have a dom node.