RSP-1861 [Docs][A11y] Manage focus when breadcrumb changes in link Popover
When a link within a popover changes the contents of the popover, focus will be lost to the document.body, because the element that had focus will be removed from the DOM and replaced with new content.  We should manage focus by explicitly setting focus to the current breadcrumb item when the content changes.

RSP-1861 Fix lint error caused by improper accessibility role for element with aria-haspopup