fix(#1478): don't tree walk to manage focus when nodeToRestore === document.body
This fixes an issue with the useFocusManager example on the documentation site,  http://localhost:1234/react-aria/FocusScope.html#usefocusmanager-example, where with an inline FocusScope, navigated into using the default keyboard navigation order for the web page, Tab or Shift+Tab does not focus the next or previous element after the FocusScope, but instead, for Tab out in the forward direction focuses the next element from the start of the document.body, and for Shift+Tab out in the reverse direction blurs the element that currently has focus, shifting focus to the document.body.

The expected behavior of Shift+Tab or Tab in this case should move focus to the next tabbable element before or after the FocusScope.