fixing docs, updating tests, fixing multiple loadMore calls in useAsyncList
found a bug where list.loadMore could be called multiple times in quick succession, resulting in a error being thrown since that case was not handled. Attempts to add a case in loadingMore for loadingMore lead to an infinite loadMore loop when user deleted all text in async Combobox (only in docs). As a result, added a ref tracker for loadingMore since data.state can have a stale value if dispatch does not finish