Speed up `yarn lint` (#1796)
This speeds up the overall performance of running `yarn lint` on most machines. Lint alone on my machine takes ~109s. TSC takes another ~30s. When run in series this is over 2min. However, when run in parallel, the entire command only takes ~109s, in otherwords only as long as the longest process.