fixing a variety of bugs found
useSelectableCollection change is to trigger a focus on the previously focused cell since we dont rerender when tabbing into the table anymore. The useSelectableItem changes are as follows: useEffect change to make pageDown/up in async tables and large tables properly render the focus ring (previously had spotty behavior). removing focus from dep array allows user to properly tab out of the table. onBlur removal is to properly retain tabindex=0 for a cell when you tab out of the table