Fix field styles when labelPosition: side
Form field and help text components (but not label) are wrapped in a vertical flex container, so that the help text renders below the field while the label is to the side. This flex container shouldn't render when no help text is provided.

Modified CSS so that this new wrapper's width defaults to the default form field width, but still respects a user override.