WIP: Use aria-required instead of required for checkboxes
This allows required, unselected checkboxes to trigger user-implemented validation instead of the native browser "Please select this checkbox" alert when it's in a form that's submitted.

For radios, repeat this fix in `useRadio` but make sure RadioGroup handles this logic. Don't do this for Switch because those components don't actually support being required in forms anyway.