fix(#2336): Fields without description are losing focus when an error is shown for the first time.
🐛 Bug Report
If you have a TextField without a description, the focus on the field is lost as soon as an error is shown for the first time. This probably happens for all the fields that have Helper Text option.

🤔 Expected Behavior
Field should keep focus.

😯 Current Behavior
Field is losing the focus when the error is shown.

💁 Possible Solution
Probably use a useLayoutEffect to restore focus to the input if the input had focus previously.

🔦 Context
💻 Code Sample
https://codesandbox.io/s/lost-focus-44wtq

🌍 Your Environment
Software	Version(s)
react-spectrum	3.14.0
Browser	Chrome
Operating System	macOS 11.6