fix(#1961): Manage focus in docs popover breadcrumbs
Fixes  #1961

When a link within a popover changes the contents of the popover, focus will be lost to the document.body, because the element that had focus will be removed from the DOM and replaced with new content.

There was prior work done on this, but we had to change some CSS to do it, we should figure out a way that doesn't require any CSS changes to the component. So this is really about managing the focus in our docs, not necessarily about Breadcrumbs themselves.
#728

We can make changes to FocusScope to better account for a content changes within the Focus resulting in a loss of focus while the scope remains active, as when the contents of a Dialog change due to actions taken within the Dialog, but for the Breadcrumbs use case, for a consistent user experience, it seems that the current BreadcrumbItem still needs to be focused explicitly.