fix(#2761): FocusScope: add allowFocusableFirstInScope prop
Add allowFocusableFirstInScope prop to specify whether to include focusable but not tabbable elements when identifying the first focusable element in scope.

This is relevant for components like a Dialog, where when the contents of the FocusScope changes, and focus gets lost, we want to restore focus to the Dialog itself, which is focusable but not tabbable.