fix(#2901): ColorArea add aria-hidden to input without focus
So that only a single "2d slider" control shows up when listing form elements for screen readers, add aria-hidden="true" to the unfocused input when the value has not changed via the keyboard, but remove aria-hidden to reveal the input for each channel when the value has changed with the keyboard.

Refactor aria-valuetext so that yChannel always comes before yChannel.