Fixing row drop indicator and make ariaHideOutside hide row and cell
Wrapped row drop indicator button in a row and cell to make it valid for accessibility tree. Made ariaHideOutside apply aria-hidden to row elements so that TalkBack will properly skip those rows. For iOS VoiceOver, we go ahead and apply aria-hidden to the cells within the row elements as well still