allowing drops that have invalid and valid drag types
this is for a case where a folder only accepts certain file types but the user drops a bunch of valid and invlid files types on it. It should still accept the drop but the invalid drag items should be filtered out by the user (for now, up for discussion). Currently there is a bug where onRemove doesnt know which of the dragged keys were valid drops