Add separate mode for triggering resizing manually via Enter (#4061)
* adding separate mode for enabling resize

this make it so the user would have to hit Enter to start resizing. This allows us to have the resizers always be visible and still preserve grid navigation between columns.

* skip test for now so build works for testing

* wrap useMove so it doesnt trigger keyboard handlers when edit mode isnt enabled

* update tableview so that it enters edit mode when choosing to resize via menu

still one bug where manager.isFocused is set to false for somereason after confirming the resize

* fixing bugs

fixes bug where selection manager would stop tracking if the manager is focused or not if keyboard nav was disabled. This was a bug where the user wouldnt be able to move table focus after confirming a resize operation. Adds new prop to useTableColumnResize for triggering resizeStart if resizer if focused. This is needed for our table since we dont have a way to call resizeStart programmatically out side of useTableColumnResize and we dont want to call it on focus for the aria example since that should require a manual trigger of Enter to enter resize mode

* updating description and missing logic

* getting rid of shouldResizeOnFocus prop

* fix skipped test

turns out the simulated resize operation was blurring before pressEnd finished

* updated docs to move inline css into classnames

also updates copy and handles case where trigger ref isnt provided