refactor GridCollection so it accepts just the top level nodes and dives into the child nodes itself
I found it really confusing that GridCollection visited the child nodes of the items provided to it but it still ran under the assumption that its items would be every row in the collection. I refactored it so you just need to pass the top level nodes (aka header row + section/rows) and it would do the rowNode adjustments appropriately by diving into the sections. This means I no longer need to pass both sections and rows to GridCollection