Fix useControlledState with suspense
We cannot write to a ref in render, because aborted renders due to suspense should not prevent later onChange events from firing. Instead of using a ref, we mutate a local variable to prevent duplicate uncontrolled onChange calls. This way it resets every render and the value and setValue functions are always in sync with each other. It does mean that the setValue function changes every time the value does now, but we should not have been relying on it to be stable because it already changed whenever the onChange function identity changed.