make NVDA/Talkback not announce the chevron label when the cell is focused
VoiceOver still announces the chevron as part of the cells content but that was true before this change anyways. Change not made in the hooks because we need to differentiate the grid cell content from the chevron but we only have gridCellProps in the hook. Make that update later when we make expandable table rows aria example