(WIP) Simplify TreeGridCollection by building the keymap inside state instead  (#4730)
* tentative update to treegridstate to minimize usage of treegrid collection

reversions of delegate and layout code to come

* revert changes to keyboard delegates and other areas

now that we are using TableCollection again with a flattened set of rows, reverted uneeded changes where possible. Kept TableLayout changes since it felt non-tree specific and is an overall improvement

* partial simplification of TreeGridCollection

one snag is that useCollection expects the collection provided as the factory to have a bunch of getters that TreeGridCollection just doesnt need

* get rid of TreeGridCollection in favor of function in useTreeGridState

TreeGridCollection was only really being used to generate a keymap of tree grid nodes to set aria-posinset and other aria attributes so it was a bit overkill and didnt need all the extra required getters, especially since TableCollection would be used in layout/keyboard delegates.