making table column header match body width with subpixel precision
clientWidth was rounding the width value of the body and setting that as the visible rect. This meant that the column header row was sometimes set a bit too wide due to the rounding up, thus causing scrollbars to appear when the table matches its flex container width. These scrollbars cause the infinite calc loop in the issue