updating menu tree state expanded keys on submenu open/close
needed a way for a menu to know that it has a open submenu so that it can set aria-hidden on itself, otherwise Safari VO had a hard time shifting focus to the newly focused submenu item. Also helpful for knowing when to set contain on a menu. Alternative would be having useMenu accept menuTreeState and initialize a level and look up in the expandedKeysStack if it owns a open submenu, open to discussion if that is the way we wanna go