fix scrollbars appearing when closing all submenus
comment for posterity: the popover container needs to be as wide as the viewport and offset to be flush with the viewport so that the popover positioning w/ flipping works. Also required to allow the menu width to grow to fit its contents but still be constrained by the max height. We remove the div when the root menu is closed so that scroll bars dont flash on to the screen because the leftOffset calculation isnt perfect due to a transform applied on left/right aligned menus that throws off the getBoundingClientRect calc