Add CollectionChildren component to improve performance
When the virtualizer updates the visible views, it triggers a re-render of the parent component (e.g. section) whose children changed. But we don't need to re-render the wrapper element, just the children. By splitting this into a separate component, we avoid re-rendering the parent when only its children changed.