fix shift tabbing in both menu and tagggroup
this fixes the logic so we always let the collection determine what to focus instead of focusing the last focused node when focusing the input. this gets rid of the dupe focus issue in taggroup while preserving the proper input focusing clearing problem for shift tabbing into a autocomplete menu. Also fixes some id assignment and other bugs like clicking on the input instead of shift tabbing then tabbing forward