feat: Automatically detect if wrapped Autocomplete collection supports virtual focus  (#8862)
* exporting contexts for quarry

* refactor autocomplete hook so it detects if it is attached to a filterable collection and if said collection supports virtual focus

* initial attempt to get rid of Dialog from S2 Popover

* fix tests

* add inner div with overflow so that popover arrow shows up

* fix items overflowing when partially scrolled into view

see https://github.com/adobe/react-spectrum/pull/7672/files

* update Popover so it accepts more values in styles instead

* fix lint

* export colorSchemeContext

* clean up extra comment

* fix chromatic

* review comments

* fix calc

* substitute custom event in favor of collection tabindex check

collection should only have a tab index if it isnt using virtual focus

* update width calculations

turns out the border of the popover shouldnt be included in the total width calculation, it is considered outside of the popover so no need to adjust. The border will be removed in favor of a box shadow to conform with update designs later

* forgot to remove commented out code

* move popover styling to user provided inner div

as per discussion with team, if someone wants to modify their popover internals, they are expected to add a inner wrapping div themselves and turn off padding like custom dialog.

* forgot to bring back styles

---------

Co-authored-by: Robert Snow <snowystinger@gmail.com>